refactor(benchmark): move reporting tasks to script
